# server-side/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CAB230: Web Computing API",
        "description": "An Express API by Shannon Stephens, N10205144",
        "version": "1.0.0"
    },
    "servers": [{
        "url": "http://localhost:3000"
    }],
...
    "paths": {
        "/register": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Registers a new user account",
                "description": "This is your starting point.",
                "operationId": "register",
                "produces": ["application/json"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "demo-user@gmail.com"
                                    },
                                    "password": {
                                        "example": "demopassword"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "json",
                                    "example": {
                                        "message": "yay! you've successfully registered your user account :)"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "json",
                                    "example": {
                                        "message": "oops! It looks like that user already exists :("
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": ["Authentication"],
                "summary": "Login with an existing user account",
                "operationId": "login",
                "produces": ["application/json"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "demo-user@gmail.com"
                                    },
                                    "password": {
                                        "example": "demopassword"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "json",
                                    "example": {
                                        "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c",
                                        "token_type": "Bearer",
                                        "expires_in": 86400
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "json",
                                    "example": {
                                        "message": "invalid login - bad password"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}